HSEARCH-3135 Move mapping annotation processors to their own file
