+++ mapper/pojo-base/src/main/java/org/hibernate/search/mapper/pojo/mapping/definition/annotation/processing/impl/IndexedEmbeddedProcessor.java
@@ -1,0 +1,63 @@
+/*
+ * Hibernate Search, full-text search for your domain model
+ *
+ * License: GNU Lesser General Public License (LGPL), version 2.1 or later
+ * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
+ */
+package org.hibernate.search.mapper.pojo.mapping.definition.annotation.processing.impl;
+
+import java.util.Collections;
+import java.util.HashSet;
+import java.util.Set;
+import java.util.stream.Stream;
+
+import org.hibernate.search.mapper.pojo.extractor.mapping.programmatic.ContainerExtractorPath;
+import org.hibernate.search.mapper.pojo.mapping.definition.annotation.IndexedEmbedded;
+import org.hibernate.search.mapper.pojo.mapping.definition.programmatic.PropertyMappingStep;
+import org.hibernate.search.mapper.pojo.model.spi.PojoPropertyModel;
+import org.hibernate.search.mapper.pojo.model.spi.PojoRawTypeModel;
+
+class IndexedEmbeddedProcessor extends PropertyAnnotationProcessor<IndexedEmbedded> {
+	IndexedEmbeddedProcessor(AnnotationProcessorHelper helper) {
+		super( helper );
+	}
+
+	@Override
+	Stream<? extends IndexedEmbedded> extractAnnotations(PojoPropertyModel<?> propertyModel) {
+		return propertyModel.getAnnotationsByType( IndexedEmbedded.class );
+	}
+
+	@Override
+	void doProcess(PropertyMappingStep mappingContext,
+			PojoRawTypeModel<?> typeModel, PojoPropertyModel<?> propertyModel,
+			IndexedEmbedded annotation) {
+		String cleanedUpPrefix = annotation.prefix();
+		if ( cleanedUpPrefix.isEmpty() ) {
+			cleanedUpPrefix = null;
+		}
+
+		Integer cleanedUpMaxDepth = annotation.maxDepth();
+		if ( cleanedUpMaxDepth.equals( -1 ) ) {
+			cleanedUpMaxDepth = null;
+		}
+
+		String[] includePathsArray = annotation.includePaths();
+		Set<String> cleanedUpIncludePaths;
+		if ( includePathsArray.length > 0 ) {
+			cleanedUpIncludePaths = new HashSet<>();
+			Collections.addAll( cleanedUpIncludePaths, includePathsArray );
+		}
+		else {
+			cleanedUpIncludePaths = Collections.emptySet();
+		}
+
+		ContainerExtractorPath extractorPath = helper.getExtractorPath( annotation.extraction() );
+
+		mappingContext.indexedEmbedded()
+				.extractors( extractorPath )
+				.prefix( cleanedUpPrefix )
+				.storage( annotation.storage() )
+				.maxDepth( cleanedUpMaxDepth )
+				.includePaths( cleanedUpIncludePaths );
+	}
+}
