+++ mapper/pojo-base/src/main/java/org/hibernate/search/mapper/pojo/mapping/definition/annotation/processing/impl/PropertyBridgeProcessor.java
@@ -1,0 +1,35 @@
+/*
+ * Hibernate Search, full-text search for your domain model
+ *
+ * License: GNU Lesser General Public License (LGPL), version 2.1 or later
+ * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
+ */
+package org.hibernate.search.mapper.pojo.mapping.definition.annotation.processing.impl;
+
+import java.lang.annotation.Annotation;
+import java.util.stream.Stream;
+
+import org.hibernate.search.mapper.pojo.bridge.mapping.annotation.declaration.PropertyBinding;
+import org.hibernate.search.mapper.pojo.bridge.mapping.programmatic.PropertyBinder;
+import org.hibernate.search.mapper.pojo.mapping.definition.programmatic.PropertyMappingStep;
+import org.hibernate.search.mapper.pojo.model.spi.PojoPropertyModel;
+import org.hibernate.search.mapper.pojo.model.spi.PojoRawTypeModel;
+
+class PropertyBridgeProcessor extends PropertyAnnotationProcessor<Annotation> {
+	PropertyBridgeProcessor(AnnotationProcessorHelper helper) {
+		super( helper );
+	}
+
+	@Override
+	Stream<? extends Annotation> extractAnnotations(PojoPropertyModel<?> propertyModel) {
+		return propertyModel.getAnnotationsByMetaAnnotationType( PropertyBinding.class );
+	}
+
+	@Override
+	void doProcess(PropertyMappingStep mappingContext,
+			PojoRawTypeModel<?> typeModel, PojoPropertyModel<?> propertyModel,
+			Annotation annotation) {
+		PropertyBinder<?> binder = helper.createPropertyBinder( annotation );
+		mappingContext.binder( binder );
+	}
+}
